refactor: separate the handlers into their own file
